pyproject.toml
[tool.poetry]
name = "yes-chef"
version = "0.1.0"
description = ""
authors = ["Daniel Ritchie <daniel@brainwavecollective.ai>"]
readme = "README.md"
packages = [{include = "yes_chef"}]

[tool.poetry.dependencies]
python = "^3.11"
pyserial = "^3.5"
pytest = "^8.3.3"
requests = "^2.32.3"
logging = "^0.4.9.6"
click = "^8.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
